fix(zfs): Check for iSCSI extents before deleting zvol
Ticket: #35757

+++ src/middlewared/middlewared/plugins/pool.py
@@ -861,6 +861,13 @@ class PoolDatasetService(CRUDService):
 
     @accepts(Str('id'))
     async def do_delete(self, id):
+        iscsi_target_extents = await self.middleware.call('iscsi.extent.query', [
+            ['type', '=', 'DISK'],
+            ['path', '=', f'zvol/{id}']
+        ])
+        if iscsi_target_extents:
+            raise CallError("This volume is in use by iSCSI extent, please remove it first.")
+
         return await self.middleware.call('zfs.dataset.delete', id)
 
     @item_method
